+++ ModelGraph/ModelGraphSTD/Graph/Runtime/Snapshot.cs
@@ -4,8 +4,8 @@ namespace ModelGraphSTD
 {
     internal class Snapshot
     {
-        private readonly List<(Edge edge, ((int x, int y)[], Facet, Facet) snapshot)> _edges;
-        private readonly List<(Node node, (int, int, byte, byte, byte, byte, Labeling, Sizing, BarWidth, FlipRotate, Aspect) snapshot)> _nodes;
+        private readonly List<(Edge edge, ((float x, float y)[], Facet, Facet) snapshot)> _edges;
+        private readonly List<(Node node, (float, float, byte, byte, byte, byte, Labeling, Sizing, BarWidth, FlipRotate, Aspect) snapshot)> _nodes;
 
         internal Snapshot(Selector selector)
         {/*
@@ -15,7 +15,7 @@ namespace ModelGraphSTD
             {
                 if (selector.Nodes.Count > 0)
                 {
-                    _nodes = new List<(Node node, (int, int, byte, byte, byte, byte, Labeling, Sizing, BarWidth, FlipRotate, Aspect))>(selector.Nodes.Count);
+                    _nodes = new List<(Node node, (float, float, byte, byte, byte, byte, Labeling, Sizing, BarWidth, FlipRotate, Aspect))>(selector.Nodes.Count);
                     foreach (var node in selector.Nodes)
                     {
                         _nodes.Add((node, node.Snapshot));
@@ -23,7 +23,7 @@ namespace ModelGraphSTD
                 }
                 if (selector.Edges.Count > 0 || selector.Points.Count > 0)
                 {
-                    _edges = new List<(Edge edge, ((int x, int y)[], Facet, Facet))>(selector.Edges.Count + selector.Points.Count);
+                    _edges = new List<(Edge edge, ((float x, float y)[], Facet, Facet))>(selector.Edges.Count + selector.Points.Count);
                     foreach (var edge in selector.Edges)
                     {
                         _edges.Add((edge, edge.Snapshot));
@@ -36,12 +36,12 @@ namespace ModelGraphSTD
             }
             else if (selector.IsNodeHit)
             {
-                _nodes = new List<(Node node, (int, int, byte, byte, byte, byte, Labeling, Sizing, BarWidth, FlipRotate, Aspect))>(1);
+                _nodes = new List<(Node node, (float, float, byte, byte, byte, byte, Labeling, Sizing, BarWidth, FlipRotate, Aspect))>(1);
                 _nodes.Add((selector.HitNode, selector.HitNode.Snapshot));
             }
             else if (selector.IsEdgeHit)
             {
-                _edges = new List<(Edge edge, ((int x, int y)[], Facet, Facet))>(1);
+                _edges = new List<(Edge edge, ((float x, float y)[], Facet, Facet))>(1);
                 _edges.Add((selector.HitEdge, selector.HitEdge.Snapshot));
             }
         }
@@ -56,7 +56,7 @@ namespace ModelGraphSTD
          */
             if (snap._nodes != null)
             {
-                _nodes = new List<(Node node, (int, int, byte, byte, byte, byte, Labeling, Sizing, BarWidth, FlipRotate, Aspect))>(snap._nodes.Count);
+                _nodes = new List<(Node node, (float, float, byte, byte, byte, byte, Labeling, Sizing, BarWidth, FlipRotate, Aspect))>(snap._nodes.Count);
                 foreach (var n in snap._nodes)
                 {
                     _nodes.Add((n.node, n.node.Snapshot));
@@ -65,7 +65,7 @@ namespace ModelGraphSTD
             }
             if (snap._edges != null)
             {
-                _edges = new List<(Edge edge, ((int x, int y)[], Facet, Facet))>(snap._edges.Count);
+                _edges = new List<(Edge edge, ((float x, float y)[], Facet, Facet))>(snap._edges.Count);
                 foreach (var e in snap._edges)
                 {
                     _edges.Add((e.edge, e.edge.Snapshot));
